+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ## Versioning
 
 This project does not follow a versioning standard. Versions are crafted after the dates; for example, the version 17.7.25 was released on July, 25th in 2017
+## xx.xx.xx
+### Fixed
+- Security issue.
+
 ## 19.7.29
 
 ### Added