+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 This project does not follow a versioning standard. Versions are crafted after the dates; for example, the version 17.7.25 was released on July, 25th in 2017
 ## xx.xx.xx
 ### Fixed
-- Security issue.
+### Added
+- Hash validation to push requests.
 
 ## 19.7.29